requirements.txt
transformers==4.49.0
accelerate>=1.10.1
datasets>=4.0.0
bitsandbytes>=0.47.0
peft==0.17.1
trl>=0.17.0
deepspeed>=0.15.4
qwen-vl-utils
timm
einops
decord
av
pillow
numpy
pandas
scipy
latex2sympy2_extended
math-verify
liger_kernel
wandb
tensorboardX
matplotlib
rich
tqdm
pycocotools
python-Levenshtein
deep-translator
openai
httpx[socks]
beautifulsoup4
torch==2.8.0
torchvision==0.23.0
flash_attn>=2.8.3